apps/poll_service/requirements.txt
Django
djangorestframework
psycopg2-binary
dotenv
pika>=1.3.0
redis
django-cors-headers
django-celery-beat